knowledge_retriever/frontapp/requirements.txt
gradio
openai
dynaconf
langchain
unstructured
spacy
numpy
google-cloud-aiplatform
